Refactor rubocop inspections to have 2 api service classes, one to build one to search
